requirements.txt
flask==3.0.2
flask-cors==4.0.0
cohere==4.48
faiss-cpu==1.7.4
numpy==1.26.4
python-dotenv==1.0.1
openai==1.12.0
langchain-community==0.0.27
gunicorn==21.2.0 